tests/wpt/web-platform-tests/referrer-policy/generic/inheritance/iframe-inheritance-srcdoc.html

<!DOCTYPE html>
<html>
  <head>
    <title>Referrer Policy: iframes srdoc correctly inherit the ancestor's referrer policy</title>
    <link rel="help" href="https://www.w3.org/TR/referrer-policy/#referrer-policy-delivery-nested">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
    <script src="resources/make-html-script.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body onload="runTest()">
    <h1>Referrer Policy: iframes srcdoc correctly inherit the ancestor's referrer policy</h1>
    <script>
      var test = async_test("iframes srcdoc correctly inherit the ancestor's referrer policy");
      window.addEventListener("message", test.step_func_done(msg => {
        assert_equals(msg.data.referrer, self.origin + "/");
      }));

      function runTest() {
        var iframe = document.createElement("iframe");
        iframe.srcdoc = createScriptString(get_host_info().REMOTE_ORIGIN);
        document.body.appendChild(iframe);
      }
    </script>
    <div id="log"></div>
  </body>
</html>
